refactor(omo): split routes/omo.py into 4 modules (Fixes #1770)
Extract 1197-line omo.py monolith into focused modules:
- services/omo_responses.py (173 lines): all Pydantic schemas + _build_upload_response
- services/omo_storage.py (91 lines): GCS helpers + _OMO_GCS_BUCKET constant
- services/omo_jobs.py (274 lines): _update_upload, _run_identification, _run_grading
- routes/omo.py (741 lines): FastAPI router + route handlers (kept #1740 lesson_id translation)

Zero behavior change: same route URLs, same response shapes, same middleware.
Test fix: update mock patch paths in test_omo_hint_1637.py to new module locations.
8 pre-existing test failures unchanged (same as staging baseline).
